refactor: combine some skeleton files and some clearer names
