+++ src/layer/segmentation/spatial_skeleton_commands.spec.ts
@@ -22,7 +22,7 @@ import {
   findSpatiallyIndexedSkeletonNode,
   getSpatiallyIndexedSkeletonDirectChildren,
   getSpatiallyIndexedSkeletonNodeParent,
-} from "#src/skeleton/edit_state.js";
+} from "#src/skeleton/node_traversal.js";
 import { SpatialSkeletonState } from "#src/skeleton/spatial_skeleton_manager.js";
 import { StatusMessage } from "#src/status.js";
