+++ src/skeleton/chunk_serialization.ts
