+++ src/skeleton/navigation_graph.ts
