+++ src/skeleton/segment_overlay.ts
@@ -14,6 +14,10 @@
  * limitations under the License.
  */
 
+/**
+ * @file GPU geometry and segment retention used for the skeleton overlay rendering pass.
+ */
+
 // Scratch buffer for GPU-upload-only arrays (segmentIds, selected, edge indices,
 // edge segmentIds). Grown monotonically; safe to reuse because SkeletonOverlayChunk
 // uploads these to the GPU synchronously and does not retain CPU references to them.
