+++ src/ui/spatial_skeleton_edit_tab.spec.ts
@@ -1,7 +1,7 @@
 import { describe, expect, it } from "vitest";
 
 import type { SpatiallyIndexedSkeletonNode } from "#src/skeleton/api.js";
-import { buildSpatiallyIndexedSkeletonNavigationGraph } from "#src/skeleton/navigation.js";
+import { buildSpatiallyIndexedSkeletonNavigationGraph } from "#src/skeleton/navigation_graph.js";
 import { SpatialSkeletonNodeFilterType } from "#src/skeleton/node_types.js";
 import { buildSpatialSkeletonSegmentRenderState } from "#src/ui/spatial_skeleton_edit_tab_render_state.js";
