+++ src/ui/spatial_skeleton_edit_tab_render_state.ts
@@ -18,7 +18,7 @@ import type { SpatiallyIndexedSkeletonNode } from "#src/skeleton/api.js";
 import {
   getFlatListNodeIds,
   type SpatiallyIndexedSkeletonNavigationGraph,
-} from "#src/skeleton/navigation.js";
+} from "#src/skeleton/navigation_graph.js";
 import {
   classifySpatialSkeletonDisplayNodeType as classifyNodeType,
   matchesSpatialSkeletonNodeFilter,
